refactor: relocate group of functions to composabels dir for clarity
