+++ packages/bridge/src/runtime/composables.ts
@@ -1,8 +1,8 @@
 export { useLazyAsyncData, refreshNuxtData } from './composables/asyncData'
 export { useLazyFetch } from './composables/fetch'
-export { useCookie } from './cookie'
+export { useCookie } from './composables/cookie'
 export { clearError, createError, isNuxtError, throwError, showError, useError } from './composables/error'
-export { useRequestHeaders, useRequestEvent } from './ssr'
+export { useRequestHeaders, useRequestEvent } from './composables/ssr'
 
 export * from 'vue'
